instructions for Creating a window and code
In the enter() function we first initialize GLFW with glfwInit(), after which we can configure GLFW using glfwWindowHint().
*0* the first arg. of glfwWindowHint 
  tells us what option we want to configure, where we can select the option from a large enum of possible options prefixed with GLFW_

*1* the second arg. 
 is an integer that sets the value of our option. A list of all the possible options and its corresponding values can be found at GLFW's window handling documentation

============================================
Next we're required to create a window object. This window object holds all the windowing data and is used quite frequently by GLFW's other functions..

==============================================
GLEW manages function pointers for OpenGL so we want to initilize GLEW before we call any OpenGL functions..
    glewExperimental = gl_true;// ensures GLEW uses more modern techniques for managing OpenGL functionality....
    if ( GLEW_OK != glewInit() ) {  return -1;   }

==============================================
We have to tell OpenGL the size of the rendering window so OpenGL knows how we want to display the data and coordinates with repect to the window. We can set those dimensions via the glViewport() function...
    glViewport(0, 0, 800, 600);